[#268] Moving connector/authorization classes to "connector" namespace.
