+++ src/Connector/HybridCredentials.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * Copyright 2018 Google Inc.
+ * Copyright 2019 Google Inc.
  *
  * This program is free software; you can redistribute it and/or modify it under
  * the terms of the GNU General Public License version 2 as published by the
@@ -17,33 +17,19 @@
  * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  */
 
-namespace Drupal\apigee_edge;
+namespace Drupal\apigee_edge\Connector;
 
+use Drupal\apigee_edge\Exception\InvalidArgumentException;
 use Drupal\apigee_edge\Plugin\EdgeKeyTypeInterface;
 use Drupal\key\KeyInterface;
-use Http\Message\Authentication;
 
 /**
- * The API credentials.
+ * The API credentials for HybridCredentials.
  */
-class Credentials implements CredentialsInterface {
+class HybridCredentials extends Credentials {
 
   /**
-   * The key entity which stores the API credentials.
-   *
-   * @var \Drupal\key\KeyInterface
-   */
-  protected $key;
-
-  /**
-   * The key type of the key entity.
-   *
-   * @var \Drupal\apigee_edge\Plugin\EdgeKeyTypeInterface
-   */
-  protected $keyType;
-
-  /**
-   * Credentials constructor.
+   * HybridCredentials constructor.
    *
    * @param \Drupal\key\KeyInterface $key
    *   The key entity which stores the API credentials.
@@ -53,33 +39,15 @@ class Credentials implements CredentialsInterface {
    *   does not implement EdgeKeyTypeInterface.
    */
   public function __construct(KeyInterface $key) {
-    if (!(($key_type = $key->getKeyType()) instanceof EdgeKeyTypeInterface)) {
-      throw new \InvalidArgumentException("Type of {$key->id()} key does not implement EdgeKeyTypeInterface.");
+    if ($key->getKeyType() instanceof EdgeKeyTypeInterface
+      && ($auth_type = $key->getKeyType()->getAuthenticationType($key))
+      && $auth_type === EdgeKeyTypeInterface::EDGE_AUTH_TYPE_JWT
+    ) {
+      parent::__construct($key);
     }
-
-    $this->key = $key;
-    $this->keyType = $key_type;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function getAuthentication(): Authentication {
-    return $this->keyType->getAuthenticationMethod($this->key);
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function getKey(): KeyInterface {
-    return $this->key;
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function getKeyType(): EdgeKeyTypeInterface {
-    return $this->keyType;
+    else {
+      throw new InvalidArgumentException("The `{$key->id()}` key is not configured for Hybrid Authentication.");
+    }
   }
 
 }
