+++ src/Connector/OauthTokenFileStorage.php
@@ -17,7 +17,7 @@
  * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  */
 
-namespace Drupal\apigee_edge;
+namespace Drupal\apigee_edge\Connector;
 
 use Drupal\apigee_edge\Exception\OauthTokenStorageException;
 use Drupal\Core\Config\ConfigFactoryInterface;
