+++ src/Connector/OauthTokenStorageInterface.php
@@ -18,24 +18,23 @@
  * MA 02110-1301, USA.
  */
 
-namespace Drupal\apigee_edge;
+namespace Drupal\apigee_edge\Connector;
 
-use Apigee\Edge\ClientInterface;
-use Apigee\Edge\HttpClient\Plugin\Authentication\Oauth;
-use Http\Message\Authentication\BasicAuth;
+use Apigee\Edge\HttpClient\Plugin\Authentication\OauthTokenStorageInterface as EdgeOauthTokenStorageInterface;
 
 /**
- * Decorator for OAuth authentication plugin.
+ * Base definition of the OAuth token storage service implementations.
  */
-class OauthAuthentication extends Oauth {
+interface OauthTokenStorageInterface extends EdgeOauthTokenStorageInterface {
 
   /**
-   * {@inheritdoc}
+   * Checks requirements of the token storage.
+   *
+   * If a requirement does not fulfilled it throws an exception.
+   *
+   * @throws \Drupal\apigee_edge\Exception\OauthTokenStorageException
+   *   Exception with the unfulfilled requirement.
    */
-  protected function authClient(): ClientInterface {
-    /** @var \Drupal\apigee_edge\SDKConnectorInterface $sdk_connector */
-    $sdk_connector = \Drupal::service('apigee_edge.sdk_connector');
-    return $sdk_connector->buildClient(new BasicAuth($this->clientId, $this->clientSecret), $this->auth_server);
-  }
+  public function checkRequirements(): void;
 
 }
