+++ src/SDKConnector.php
@@ -23,6 +23,9 @@ use Apigee\Edge\Api\Management\Controller\OrganizationController;
 use Apigee\Edge\Client;
 use Apigee\Edge\ClientInterface;
 use Apigee\Edge\HttpClient\Utility\Builder;
+use Drupal\apigee_edge\Connector\CredentialsInterface;
+use Drupal\apigee_edge\Connector\HybridCredentials;
+use Drupal\apigee_edge\Connector\OauthCredentials;
 use Drupal\apigee_edge\Exception\AuthenticationKeyException;
 use Drupal\apigee_edge\Exception\AuthenticationKeyNotFoundException;
 use Drupal\apigee_edge\Exception\InvalidArgumentException;
@@ -52,7 +55,7 @@ class SDKConnector implements SDKConnectorInterface {
   /**
    * The currently used credentials object.
    *
-   * @var null|\Drupal\apigee_edge\CredentialsInterface
+   * @var null|\Drupal\apigee_edge\Connector\CredentialsInterface
    */
   private static $credentials = NULL;
 
@@ -189,7 +192,7 @@ class SDKConnector implements SDKConnectorInterface {
   /**
    * Returns the credentials object used by the API client.
    *
-   * @return \Drupal\apigee_edge\CredentialsInterface
+   * @return \Drupal\apigee_edge\Connector\CredentialsInterface
    *   The key entity.
    */
   private function getCredentials(): CredentialsInterface {
@@ -210,7 +213,7 @@ class SDKConnector implements SDKConnectorInterface {
   /**
    * Changes credentials used by the API client.
    *
-   * @param \Drupal\apigee_edge\CredentialsInterface $credentials
+   * @param \Drupal\apigee_edge\Connector\CredentialsInterface $credentials
    *   The new credentials object.
    */
   private function setCredentials(CredentialsInterface $credentials) {
@@ -225,7 +228,7 @@ class SDKConnector implements SDKConnectorInterface {
    * @param \Drupal\key\KeyInterface $key
    *   The key entity which stores the API credentials.
    *
-   * @return \Drupal\apigee_edge\CredentialsInterface
+   * @return \Drupal\apigee_edge\Connector\CredentialsInterface
    *   The credentials.
    */
   private function buildCredentials(KeyInterface $key): CredentialsInterface {
